Implement DEFINE_STATICCLASS, GetDefaultObj, GetOutermost, and fix CrashReporter bugs
- Add DEFINE_STATICCLASS macro to PropertyMarcos.h with StaticClass() and GetDefaultObj().
- Add defensive nullptr check in GetDefaultObj().
- Define UObject, UClass, and UPackage in UObjectBase.h.
- Add GetName, GetPathName, GetFullName, and GetOutermost helpers to UObject.
- Implement recursive path name resolution and class default object finding in UObjectArray.cpp.
- Add overloaded StaticFindObject support for wchar_t and char paths.
- Fix FName::ToString() bug to correctly invoke the original ToStringOG function pointer.
- Fix CrashReporter target process termination upon crash to release DLL file locks and resolve LNK1104.
- Fix CrashReporter stack trace generation bug by preventing OverflowException when converting IntPtr to ToInt32 on 64-bit systems.

+++ FortRuntime/CrashReporter.cs
@@ -593,8 +593,8 @@ namespace FortRuntime
 
                                     // 3. Write EXCEPTION_POINTERS (ExceptionRecord, ContextRecord)
                                     byte[] pointersBytes = new byte[8];
-                                    Array.Copy(BitConverter.GetBytes(remoteExceptionRecord.ToInt32()), 0, pointersBytes, 0, 4);
-                                    Array.Copy(BitConverter.GetBytes(remoteContext.ToInt32()), 0, pointersBytes, 4, 4);
+                                    Array.Copy(BitConverter.GetBytes((uint)remoteExceptionRecord.ToInt64()), 0, pointersBytes, 0, 4);
+                                    Array.Copy(BitConverter.GetBytes((uint)remoteContext.ToInt64()), 0, pointersBytes, 4, 4);
                                     WriteProcessMemory(hProcess, remoteExceptionPointers, pointersBytes, 8, out _);
 
                                     // Generate perfect UE4Minidump.dmp with remote pointers
